requirements.txt
requests>=2.23.0
tqdm>=4.43.0
pandas>=1.0.1
mysql-connector-python>=8.0.19
bs4>=0.0.1
